requirements-dev.txt
# Dependencies for running tests and development scripts
requests
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
httpx>=0.24.0
respx>=0.20.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
google-cloud-aiplatform>=1.25.0
google-adk
fastapi>=0.100.0
uvicorn>=0.20.0 